+++ src/api/controllers/tickersearchcontroller.js
@@ -1,0 +1,5 @@
+const tickersearchservice = require("../services/tickersearchservice");
+
+module.exports = function (router) {
+    router.get('/tickersearch', tickersearchservice.getTickerSearch);
+};
