+++ src/api/utils/db.js
@@ -19,4 +19,5 @@ module.exports = {
   User: require("../models/user"),
   CompanyOverview: require("../models/companyoverview"),
   NewsAndSentiments: require("../models/newsandsentiments"),
+  TickerSearch: require("../models/tickerseach"),
 };
